+++ Services/API/IProductService.cs
@@ -6,9 +6,9 @@ public interface IProductService
 {
     IEnumerable<IProductData> GetAllProducts();
     IProductData GetProduct(int productId);
-    bool AddProduct(int productId, string name, float price);
-    bool AddProduct(int productId, string name, string description, float price);
-    bool UpdateProduct(int productId, string name, string description, float price);
+    bool AddProduct(int productId, string name, int price);
+    bool AddProduct(int productId, string name, string description, int price);
+    bool UpdateProduct(int productId, string name, string description, int price);
     bool DeleteProduct(int productId);
     void NukeData();
 }
